move repo data to its own directory, add readme
